pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nhi-agent"
version = "0.1.0"
description = "Identity management and analysis tool that interacts with AWS and Vault MCP servers"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.34.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.12",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
nhi-agent = "src.main:app"

[tool.setuptools]
packages = ["src"]





